# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: |-
    REST API for managing listeners, reverse-shell sessions, file transfer, and tunnels.
    Every endpoint except /api/v1/auth/token requires a Bearer token obtained via that endpoint.
  title: Platypus API
  version: "1.0"
paths:
  /api/client:
    get:
      description: Returns every session (plain + termite) across every listener.
        This is the main session-list endpoint used by the desktop UI.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all sessions
      tags:
      - sessions
  /api/client/{hash}:
    delete:
      parameters:
      - description: Session hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete session
      tags:
      - sessions
    get:
      parameters:
      - description: Session hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get session
      tags:
      - sessions
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Runs a single command on one session and returns its stdout. Submit
        cmd as form-encoded. Plain TCP sessions in PTY mode are refused; switch to
        non-PTY or use the WebSocket terminal instead.
      parameters:
      - description: Session hash
        in: path
        name: hash
        required: true
        type: string
      - description: Shell command
        in: formData
        name: cmd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Execute command
      tags:
      - sessions
  /api/client/{hash}/upgrade/{target}:
    get:
      description: Compile a Termite agent matching the client's architecture, push
        it over the existing plain shell, and let it reconnect to the target encrypted
        listener. Progress is broadcast on /notify; this endpoint only acknowledges
        the request.
      parameters:
      - description: Source session hash (plain TCP client)
        in: path
        name: hash
        required: true
        type: string
      - description: Destination listener hash (must be encrypted)
        in: path
        name: target
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upgrade to Termite
      tags:
      - sessions
  /api/server:
    get:
      description: Returns every configured TCP listener plus the distributor snapshot.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List listeners
      tags:
      - listeners
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Opens a new reverse-shell listener. Submit as application/x-www-form-urlencoded.
      parameters:
      - description: Bind address (e.g. 0.0.0.0)
        in: formData
        name: host
        required: true
        type: string
      - description: Port 1-65535
        in: formData
        name: port
        required: true
        type: integer
      - description: true for Termite listener, false for plain reverse shell
        in: formData
        name: encrypted
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create listener
      tags:
      - listeners
  /api/server/{hash}:
    delete:
      parameters:
      - description: Listener hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete listener
      tags:
      - listeners
    get:
      parameters:
      - description: Listener hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get listener
      tags:
      - listeners
  /api/server/{hash}/client:
    get:
      parameters:
      - description: Listener hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List sessions on a listener
      tags:
      - listeners
...
securityDefinitions:
  BearerAuth:
    description: Value should be "Bearer <token>". Fetch a token via POST /api/v1/auth/token
      using the secret printed at server startup.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"